move debug logic to seperate package
